--- packages/twenty-server/src/engine/metadata-modules/remote-server/remote-table/utils/excluded-postgres-schemas.ts
@@ -1,6 +1,0 @@
-export const EXCLUDED_POSTGRES_SCHEMAS = [
-  'information_schema',
-  'pg_catalog',
-  'pg_toast',
-  'public',
-];
